requirements.txt
aiohttp==3.6.2
python-dotenv==0.11.0
opencv-python==4.2.0.32
